+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "verbatimModuleSyntax": true,
     "target": "ESNext",
     "lib": ["esnext", "dom", "DOM.Iterable"],
     "module": "ESNext",
